Added macro for hook and wrapping function definition.
