+++ module/hook.c
@@ -4,78 +4,13 @@
  */
 
 #include "tracer.h"
+#include "hook.h"
 #include <linux/ftrace.h>
 #include <linux/kallsyms.h>
 #include <linux/msg.h>
 #include <linux/slab.h>
-#include <linux/syscalls.h>
-
-struct fm_hook_metadata {
-	const char *name;
-	const void *func;
-	const char *rtype;
-	const char **types;
-	const char **args;
-	struct list_head list;
-};
 
 LIST_HEAD(fm_hook_list);
-
-#define __FMH_STR_ADECL(t, a)	#a
-#define __FMH_STR_TDECL(t, a)	#t
-#define __FMH_STR_TRET (t)		#t
-#define __MAPR(m, t)			m(t)
-
-#define FM_HOOK_FUNC_DEFINE1(name, ...) FM_HOOK_FUNC_DEFINEx(1, _##name, __VA_ARGS__)
-#define FM_HOOK_FUNC_DEFINE2(name, ...) FM_HOOK_FUNC_DEFINEx(2, _##name, __VA_ARGS__)
-#define FM_HOOK_FUNC_DEFINE3(name, ...) FM_HOOK_FUNC_DEFINEx(3, _##name, __VA_ARGS__)
-#define FM_HOOK_FUNC_DEFINE4(name, ...) FM_HOOK_FUNC_DEFINEx(4, _##name, __VA_ARGS__)
-#define FM_HOOK_FUNC_DEFINE5(name, ...) FM_HOOK_FUNC_DEFINEx(5, _##name, __VA_ARGS__)
-#define FM_HOOK_FUNC_DEFINE6(name, ...) FM_HOOK_FUNC_DEFINEx(6, _##name, __VA_ARGS__)
-
-#define FM_HOOK_FUNC_DEFINE_MAXARGS  6
-#define FM_HOOK_WRAP_DEFINE_MAXARGS  6
-
-#define FM_HOOK_FUNC_DEFINEx(x, sname, ...)				\
-	__FM_HOOK_FUNC_DEFINEx(x, sname, __VA_ARGS__)		\
-	__FM_HOOK_WRAP_DEFINEx(x, sname, __VA_ARGS__)		\
-	__FM_HOOK_META_DEFINEx(x, sname, __VA_ARGS__)
-
-#define __FM_HOOK_FUNC_DEFINEx(x, name, rtype, ...)		\
-	rtype (*__do_fm_hook##name)(__MAP(x,__SC_DECL,__VA_ARGS__));
-
-#define __FM_HOOK_WRAP_DEFINEx(x, name, rtype, ...)		\
-    rtype (__do_fm_wrap##name)(__MAP(x,__SC_DECL,__VA_ARGS__)) {	\
-		rtype t;													\
-		pr_info("+ "#name);							\
-		t = __do_fm_hook##name(__MAP(x,__SC_ARGS,__VA_ARGS__));		\
-		return t; 													\
-	}
-
-#define __FM_HOOK_META_DEFINEx(x, sname, srtype, ...) 	\
-    static const char *types##sname[] = {				\
-        __MAP(x,__FMH_STR_TDECL,__VA_ARGS__)			\
-    };													\
-    static const char *args##sname[] = {				\
-        __MAP(x,__FMH_STR_ADECL,__VA_ARGS__)			\
-    };													\
-	static struct fm_hook_metadata __used				\
-	 __fm_hook_meta##sname = {							\
-		.name 	= #sname,								\
-		.func 	= __do_fm_wrap##sname,					\
-		.rtype	= #srtype, 								\
-  		.types 	= x ? types##sname : NULL,				\
-  		.args	= x ? args##sname : NULL,				\
-  		.list 	= LIST_HEAD_INIT(__fm_hook_meta##sname.list),  \
-	};													\
-  	list_add(&__fm_hook_meta##name.list, &fm_hook_list);
-
-
-#define FM_HOOK_FUNC_NAME(name)		__do_fm_hook_##name
-#define FM_HOOK_WRAP_NAME(name)		__do_fm_wrap_##name
-
-
-FM_HOOK_FUNC_DEFINE2(load_msg, struct msg_msg *, const void __user *, src, size_t, len);
 
 static void notrace hook_callback(unsigned long ip, unsigned long parent_ip,
 	struct ftrace_ops *ops, struct pt_regs *regs);
@@ -88,16 +23,12 @@ static struct ftrace_ops ops __read_mostly = {
 				| FTRACE_OPS_FL_IPMODIFY
 };
 
-// Wrapper to load_msg
-struct msg_msg *wr_load_msg(const void __user *src, size_t len)
+FM_HOOK_FUNC_DEFINE2(load_msg, struct msg_msg *, const void __user *, src, size_t, len)
 {
-	struct msg_msg *msg;
-
-	pr_info("+ load_msg()\n");
-	msg = (*__do_fm_hook_load_msg)(src, len);
-	pr_info("result: %p\n", msg);
-
-	return msg;
+	struct msg_msg *t;
+	pr_info("+ %s", "load_msg");
+	t = FM_HOOK_FUNC_NAME(load_msg)(src, len);
+	return t;
 }
 
 static void notrace hook_callback(unsigned long ip, unsigned long parent_ip,
